refactor: Move bid page logic
